+++ calcipy/doit_tasks/__init__.py
@@ -2,7 +2,7 @@
 
 Register all defaults DoIt tasks in a dodo.py file with the below snippet:
 
-`from calcipy.doit_helpers import *  # noqa: F401,F403,H303 (Run 'doit list' to see tasks)`
+`from calcipy.doit_tasks import *  # noqa: F401,F403,H303 (Run 'doit list' to see tasks)`
 
 """
