+++ calcipy/doit_tasks/doit_globals.py
