+++ calcipy/tag_collector.py
@@ -11,8 +11,8 @@ import attr
 from loguru import logger
 
 from . import __pkg_name__
-from .doit_helpers.base import debug_task, read_lines
-from .doit_helpers.doit_globals import DIG, DoItTask
+from .doit_tasks.base import debug_task, read_lines
+from .doit_tasks.doit_globals import DIG, DoItTask
 from .log_helpers import log_fun
 
 _TAG_SUMMARY_FILENAME = 'TAG_SUMMARY.md'
