+++ dodo.py
@@ -23,9 +23,9 @@ from pathlib import Path
 from loguru import logger
 
 from calcipy import LOGGER_CONFIG
-from calcipy.doit_helpers.base import debug_task
-from calcipy.doit_helpers.doit_globals import DIG, DoItTask
-from calcipy.doit_helpers import *  # noqa: F401,F403,H303 skipcq: PYL-W0614 (Run 'doit list' to see tasks)
+from calcipy.doit_tasks.base import debug_task
+from calcipy.doit_tasks.doit_globals import DIG, DoItTask
+from calcipy.doit_tasks import *  # noqa: F401,F403,H303 skipcq: PYL-W0614 (Run 'doit list' to see tasks)
 
 logger.configure(**LOGGER_CONFIG)
 logger.info('Starting DoIt tasks in dodo.py')
