--- tests/test_doit_helpers/__init__.py
@@ -1,1 +1,0 @@
-"""Test calcipy/doit_helpers."""
