+++ tests/test_doit_tasks/__init__.py
@@ -1,0 +1,1 @@
+"""Test calcipy/doit_tasks."""
