+++ tests/test_doit_tasks/test_test.py
@@ -1,9 +1,9 @@
-"""Test doit_helpers/test.py."""
+"""Test doit_tasks/test.py."""
 
 import pytest
 
-from calcipy.doit_helpers.doit_globals import DIG
-from calcipy.doit_helpers.test import task_test_marker
+from calcipy.doit_tasks.doit_globals import DIG
+from calcipy.doit_tasks.test import task_test_marker
 
 from ..configuration import DIG_CWD
